Can force options with config.php

+++ plugin/Admin.php
@@ -6,25 +6,6 @@
 class Admin {
 
 	/**
-	 * @var bool Whether options should be taken from config file
-	 */
-	private $readonly;
-
-	/**
-	 * Handles options via config file
-	 */
-	public function __construct() {
-		$this->readonly = ( file_exists(__DIR__ . '/Config.php' ) ) ? true : false;
-		if ( $this->readonly ) {
-			$config = require 'Config.php';
-			update_option( 'marketo_pro_client_id',     $config['client_id'] );
-			update_option( 'marketo_pro_client_secret', $config['client_secret'] );
-			update_option( 'marketo_pro_munchkin_id',   $config['munchkin_id'] );
-			update_option( 'marketo_pro_marketo_id',    $config['marketo_id'] );
-		}
-	}
-
-	/**
 	 * Called by admin menu hook
 	 */
 	public function add_menus() {
@@ -36,8 +17,9 @@ class Admin {
 	 * Bootstraps the admin page, called by `add_menu_page`
 	 */
 	public function admin() {
-		$data['readonly']     = get_option('marketo_pro_client_id');
+		global $marketo_pro_readonly;
 
+		$data['readonly']      = $marketo_pro_readonly;
 		$data['client_id']     = get_option('marketo_pro_client_id');
 		$data['client_secret'] = get_option('marketo_pro_client_secret');
 		$data['munchkin_id']   = get_option('marketo_pro_munchkin_id');
@@ -47,22 +29,26 @@ class Admin {
 	}
 
 	/**
-	 * Escapes and stores options
+	 * Escapes and stores options if not in readonly mode
 	 *
 	 * @param array $post $_POST data
 	 */
 	public function save_options( $post ) {
-		$options['client_id']     = $post['client_id'];
-		$options['client_secret'] = $post['client_secret'];
-		$options['munchkin_id']   = $post['munchkin_id'];
-		$options['marketo_id']    = $post['marketo_id'];
+		global $marketo_pro_readonly;
 
-		foreach ($options as $key => $value) $options[ $key ] = preg_replace( '/[^\w-]/', '', $value );
+		if ( !$marketo_pro_readonly ) {
+			$options['client_id']     = $post['client_id'];
+			$options['client_secret'] = $post['client_secret'];
+			$options['munchkin_id']   = $post['munchkin_id'];
+			$options['marketo_id']    = $post['marketo_id'];
 
-		update_option( 'marketo_pro_client_id',     $options['client_id'] );
-		update_option( 'marketo_pro_client_secret', $options['client_secret'] );
-		update_option( 'marketo_pro_munchkin_id',   $options['munchkin_id'] );
-		update_option( 'marketo_pro_marketo_id',    $options['marketo_id'] );
+			foreach ($options as $key => $value) $options[ $key ] = preg_replace( '/[^\w-]/', '', $value );
+
+			update_option( 'marketo_pro_client_id',     $options['client_id'] );
+			update_option( 'marketo_pro_client_secret', $options['client_secret'] );
+			update_option( 'marketo_pro_munchkin_id',   $options['munchkin_id'] );
+			update_option( 'marketo_pro_marketo_id',    $options['marketo_id'] );
+		}
 
 		wp_redirect( 'admin.php?page=marketo-pro&updated=true' );
 	}
